requirements.txt
requests
ics
pytz
aiogram
python-dotenv
psycopg2-binary